% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel.demographics}
\alias{channel.demographics}
\title{Query the Analytics API for channel demographics metrics}
\usage{
channel.demographics(
  channel_id,
  start_date = NULL,
  end_date = NULL,
  filters = NULL
)
}
\arguments{
\item{channel_id}{YouTube channel ID}

\item{start_date}{Beginning date of the query date range. YYYY-mm-dd}

\item{end_date}{End date of the query date range. YYYY-mm-dd}
}
\value{
data.frame
}
\description{
Query the Analytics API for channel demographics metrics
}
