% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel.query}
\alias{channel.query}
\title{Query the analytics API for channel metrics}
\usage{
channel.query(
  channel_id,
  start_date = NULL,
  end_date = NULL,
  metrics = NULL,
  dimensions = NULL,
  sort = NULL,
  filters = NULL,
  max_results = NULL
)
}
\arguments{
\item{channel_id}{YouTube channel ID string}

\item{start_date}{Beginning date of the query date range}

\item{end_date}{End date of the query date range}

\item{metrics}{Comma-seperated string of metric names}

\item{sort}{Comma-seperated string of dimension names to sort by.  Prepend
dimension names with '-' to sort that dimension in descending order}

\item{filters}{Semi-colon seperated string of filter conditions to apply
to the result set.}

\item{max_results}{Maximum number of results to return. (default 1000)}

\item{dimensiosn}{Comma-seperated string of dimension names}
}
\value{
data.frame containing the API response rows
}
\description{
Query the analytics API for channel metrics
}
