% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{Empty}
\alias{Empty}
\title{YouTube Reporting API Objects
Schedules reporting jobs containing your YouTube Analytics data and downloads the resulting bulk data reports in the form of CSV files.}
\usage{
Empty()
}
\value{
Empty object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2022-01-14 13:42:58
filename: /Users/jalvarado/apps//googleyoutubereportingv1.auto/R/youtubereporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Empty Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for \code{Empty} is empty JSON object \code{{}}.
}
\concept{Empty functions}
