% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{EmptyResponse}
\alias{EmptyResponse}
\title{EmptyResponse Object}
\usage{
EmptyResponse(errors = NULL)
}
\arguments{
\item{errors}{Apiary error details}
}
\value{
EmptyResponse object
}
\description{
EmptyResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Empty response.
}
\concept{EmptyResponse functions}
