% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{ErrorProto}
\alias{ErrorProto}
\title{ErrorProto Object}
\usage{
ErrorProto(
  location = NULL,
  domain = NULL,
  externalErrorMessage = NULL,
  argument = NULL,
  code = NULL,
  locationType = NULL,
  debugInfo = NULL
)
}
\arguments{
\item{location}{Location of the error, as specified by the location type}

\item{domain}{Error domain}

\item{externalErrorMessage}{A short explanation for the error, which can be shared outside Google}

\item{argument}{Error arguments, to be used when building user-friendly error messages given the error domain and code}

\item{code}{Error code in the error domain}

\item{locationType}{No description}

\item{debugInfo}{Debugging information, which should not be shared externally}
}
\value{
ErrorProto object
}
\description{
ErrorProto Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes one specific error.
}
\concept{ErrorProto functions}
