% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{Errors}
\alias{Errors}
\title{Errors Object}
\usage{
Errors(requestId = NULL, error = NULL, code = NULL)
}
\arguments{
\item{requestId}{Request identifier generated by the service, which can be used to identify the error in the logs}

\item{error}{Specific error description and codes}

\item{code}{Global error code}
}
\value{
Errors object
}
\description{
Errors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request Error information. The presence of an error field signals that the operation has failed.
}
\concept{Errors functions}
