% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{find_playlists}
\alias{find_playlists}
\title{Search for playlists matching the given criteria.}
\usage{
find_playlists(channel_id = NULL, title = NULL, q = NULL, ...)
}
\arguments{
\item{channel_id}{Character containing the YouTube channel ID. Used to limit
the search results to only playlists belonging to the specified channel.}

\item{title}{Character containing a playlist title to search for.  Will
only return exact matches and is case-sensitive.  Either \code{title} or
\code{q} parameter is retuired, but not both.}

\item{q}{Search term used to filter playlists by title.  Takes any regular
expression supported by \code{grepl}. Optional.}

\item{...}{Additional parameters passed to \code{tuber::get_playlists}}
}
\description{
Search for playlists matching the given criteria.
}
\seealso{
\code{\link[tuber:get_playlists]{tuber::get_playlists()}} for additional parameters

\code{\link[=grepl]{grepl()}} for regular expression requirements in the \code{q}
parameter
}
