% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataBlobstore2Info}
\alias{GdataBlobstore2Info}
\title{GdataBlobstore2Info Object}
\usage{
GdataBlobstore2Info(
  readToken = NULL,
  blobGeneration = NULL,
  downloadReadHandle = NULL,
  uploadMetadataContainer = NULL,
  blobId = NULL
)
}
\arguments{
\item{readToken}{gdata}

\item{blobGeneration}{gdata}

\item{downloadReadHandle}{gdata}

\item{uploadMetadataContainer}{gdata}

\item{blobId}{gdata}
}
\value{
GdataBlobstore2Info object
}
\description{
GdataBlobstore2Info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataBlobstore2Info functions}
