% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataCompositeMedia}
\alias{GdataCompositeMedia}
\title{GdataCompositeMedia Object}
\usage{
GdataCompositeMedia(
  objectId = NULL,
  referenceType = NULL,
  path = NULL,
  blobstore2Info = NULL,
  sha1Hash = NULL,
  md5Hash = NULL,
  inline = NULL,
  length = NULL,
  crc32cHash = NULL,
  cosmoBinaryReference = NULL,
  blobRef = NULL
)
}
\arguments{
\item{objectId}{gdata}

\item{referenceType}{gdata}

\item{path}{gdata}

\item{blobstore2Info}{gdata}

\item{sha1Hash}{gdata}

\item{md5Hash}{gdata}

\item{inline}{gdata}

\item{length}{gdata}

\item{crc32cHash}{gdata}

\item{cosmoBinaryReference}{gdata}

\item{blobRef}{gdata}
}
\value{
GdataCompositeMedia object
}
\description{
GdataCompositeMedia Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataCompositeMedia functions}
