% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataContentTypeInfo}
\alias{GdataContentTypeInfo}
\title{GdataContentTypeInfo Object}
\usage{
GdataContentTypeInfo(
  fromBytes = NULL,
  bestGuess = NULL,
  fromFileName = NULL,
  fromHeader = NULL,
  fromUrlPath = NULL
)
}
\arguments{
\item{fromBytes}{gdata}

\item{bestGuess}{gdata}

\item{fromFileName}{gdata}

\item{fromHeader}{gdata}

\item{fromUrlPath}{gdata}
}
\value{
GdataContentTypeInfo object
}
\description{
GdataContentTypeInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataContentTypeInfo functions}
