% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataDiffChecksumsResponse}
\alias{GdataDiffChecksumsResponse}
\title{GdataDiffChecksumsResponse Object}
\usage{
GdataDiffChecksumsResponse(
  objectSizeBytes = NULL,
  checksumsLocation = NULL,
  objectVersion = NULL,
  objectLocation = NULL,
  chunkSizeBytes = NULL
)
}
\arguments{
\item{objectSizeBytes}{gdata}

\item{checksumsLocation}{gdata}

\item{objectVersion}{gdata}

\item{objectLocation}{gdata}

\item{chunkSizeBytes}{gdata}
}
\value{
GdataDiffChecksumsResponse object
}
\description{
GdataDiffChecksumsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataDiffChecksumsResponse functions}
