% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataDiffDownloadResponse}
\alias{GdataDiffDownloadResponse}
\title{GdataDiffDownloadResponse Object}
\usage{
GdataDiffDownloadResponse(objectLocation = NULL)
}
\arguments{
\item{objectLocation}{gdata}
}
\value{
GdataDiffDownloadResponse object
}
\description{
GdataDiffDownloadResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataDiffDownloadResponse functions}
