% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataDiffUploadRequest}
\alias{GdataDiffUploadRequest}
\title{GdataDiffUploadRequest Object}
\usage{
GdataDiffUploadRequest(
  objectVersion = NULL,
  objectInfo = NULL,
  checksumsInfo = NULL
)
}
\arguments{
\item{objectVersion}{gdata}

\item{objectInfo}{gdata}

\item{checksumsInfo}{gdata}
}
\value{
GdataDiffUploadRequest object
}
\description{
GdataDiffUploadRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataDiffUploadRequest functions}
