% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataDiffUploadResponse}
\alias{GdataDiffUploadResponse}
\title{GdataDiffUploadResponse Object}
\usage{
GdataDiffUploadResponse(originalObject = NULL, objectVersion = NULL)
}
\arguments{
\item{originalObject}{gdata}

\item{objectVersion}{gdata}
}
\value{
GdataDiffUploadResponse object
}
\description{
GdataDiffUploadResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataDiffUploadResponse functions}
