% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataDiffVersionResponse}
\alias{GdataDiffVersionResponse}
\title{GdataDiffVersionResponse Object}
\usage{
GdataDiffVersionResponse(objectSizeBytes = NULL, objectVersion = NULL)
}
\arguments{
\item{objectSizeBytes}{gdata}

\item{objectVersion}{gdata}
}
\value{
GdataDiffVersionResponse object
}
\description{
GdataDiffVersionResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataDiffVersionResponse functions}
