% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataDownloadParameters}
\alias{GdataDownloadParameters}
\title{GdataDownloadParameters Object}
\usage{
GdataDownloadParameters(allowGzipCompression = NULL, ignoreRange = NULL)
}
\arguments{
\item{allowGzipCompression}{gdata}

\item{ignoreRange}{gdata}
}
\value{
GdataDownloadParameters object
}
\description{
GdataDownloadParameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataDownloadParameters functions}
