% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataMedia}
\alias{GdataMedia}
\title{GdataMedia Object}
\usage{
GdataMedia(
  cosmoBinaryReference = NULL,
  diffUploadRequest = NULL,
  md5Hash = NULL,
  mediaId = NULL,
  sha256Hash = NULL,
  timestamp = NULL,
  isPotentialRetry = NULL,
  bigstoreObjectRef = NULL,
  inline = NULL,
  diffDownloadResponse = NULL,
  contentTypeInfo = NULL,
  length = NULL,
  compositeMedia = NULL,
  sha1Hash = NULL,
  diffVersionResponse = NULL,
  objectId = NULL,
  token = NULL,
  crc32cHash = NULL,
  path = NULL,
  filename = NULL,
  hash = NULL,
  referenceType = NULL,
  hashVerified = NULL,
  blobstore2Info = NULL,
  downloadParameters = NULL,
  algorithm = NULL,
  contentType = NULL,
  blobRef = NULL,
  diffUploadResponse = NULL,
  diffChecksumsResponse = NULL
)
}
\arguments{
\item{cosmoBinaryReference}{gdata}

\item{diffUploadRequest}{gdata}

\item{md5Hash}{gdata}

\item{mediaId}{gdata}

\item{sha256Hash}{gdata}

\item{timestamp}{gdata}

\item{isPotentialRetry}{gdata}

\item{bigstoreObjectRef}{gdata}

\item{inline}{gdata}

\item{diffDownloadResponse}{gdata}

\item{contentTypeInfo}{gdata}

\item{length}{gdata}

\item{compositeMedia}{gdata}

\item{sha1Hash}{gdata}

\item{diffVersionResponse}{gdata}

\item{objectId}{gdata}

\item{token}{gdata}

\item{crc32cHash}{gdata}

\item{path}{gdata}

\item{filename}{gdata}

\item{hash}{gdata}

\item{referenceType}{gdata}

\item{hashVerified}{gdata}

\item{blobstore2Info}{gdata}

\item{downloadParameters}{gdata}

\item{algorithm}{gdata}

\item{contentType}{gdata}

\item{blobRef}{gdata}

\item{diffUploadResponse}{gdata}

\item{diffChecksumsResponse}{gdata}
}
\value{
GdataMedia object
}
\description{
GdataMedia Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataMedia functions}
