% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{GdataObjectId}
\alias{GdataObjectId}
\title{GdataObjectId Object}
\usage{
GdataObjectId(bucketName = NULL, generation = NULL, objectName = NULL)
}
\arguments{
\item{bucketName}{gdata}

\item{generation}{gdata}

\item{objectName}{gdata}
}
\value{
GdataObjectId object
}
\description{
GdataObjectId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
gdata
}
\concept{GdataObjectId functions}
