% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_details.R
\name{get_playlist_details}
\alias{get_playlist_details}
\title{Retrieve playlist information}
\usage{
get_playlist_details(playlist_id, part = "snippet", ...)
}
\arguments{
\item{playlist_id}{character; YouTube playlist ID used to uniquely identify
a playlist.}

\item{part}{Required.  One of the following: contentDetails, id,
localizations, player, snippet, status.  Default: snippet.}

\item{...}{Additional parameters passed to \link{tuber::get_playlists}}
}
\value{
list
}
\description{
Retrieve playlist information
}
