% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_details.R
\name{get_video_details}
\alias{get_video_details}
\title{Fetch video details from the YouTube Data API}
\usage{
get_video_details(video_id = NULL, part = "snippet", ...)
}
\arguments{
\item{video_id}{Comma separated list of IDs of the videos for which
details are requested. Required.}

\item{part}{Comma-separated vector of video resource properties requested.
Options include:
\code{contentDetails, fileDetails, id, liveStreamingDetails,
localizations, player, processingDetails,
recordingDetails, snippet, statistics, status, suggestions, topicDetails}}

\item{...}{Additional arguments passed to \code{\link[tuber]{tuber_GET}}.}
}
\value{
list
}
\description{
Fetch video details from the YouTube Data API
}
