% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{Group}
\alias{Group}
\title{Group Object}
\usage{
Group(
  etag = NULL,
  snippet = NULL,
  id = NULL,
  errors = NULL,
  contentDetails = NULL
)
}
\arguments{
\item{etag}{The Etag of this resource}

\item{snippet}{The \code{snippet} object contains basic information about the group, including its creation date and name}

\item{id}{The ID that YouTube uses to uniquely identify the group}

\item{errors}{Apiary error details}

\item{contentDetails}{The \code{contentDetails} object contains additional information about the group, such as the number and type of items that it contains}
}
\value{
Group object
}
\description{
Group Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group.
}
\seealso{
Other Group functions: 
\code{\link{groups.insert}()},
\code{\link{groups.update}()}
}
\concept{Group functions}
