% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupContentDetails}
\alias{GroupContentDetails}
\title{GroupContentDetails Object}
\usage{
GroupContentDetails(itemCount = NULL, itemType = NULL)
}
\arguments{
\item{itemCount}{The number of items in the group}

\item{itemType}{The type of resources that the group contains}
}
\value{
GroupContentDetails object
}
\description{
GroupContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group's content details.
}
\concept{GroupContentDetails functions}
