% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupItem}
\alias{GroupItem}
\title{GroupItem Object}
\usage{
GroupItem(
  resource = NULL,
  groupId = NULL,
  errors = NULL,
  etag = NULL,
  id = NULL
)
}
\arguments{
\item{resource}{The \code{resource} object contains information that identifies the item being added to the group}

\item{groupId}{The ID that YouTube uses to uniquely identify the group that contains the item}

\item{errors}{Apiary error details}

\item{etag}{The Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the \code{channel}, \code{video}, \code{playlist}, or \code{asset} resource that is included in the group}
}
\value{
GroupItem object
}
\description{
GroupItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group item.
}
\seealso{
Other GroupItem functions: 
\code{\link{groupItems.insert}()}
}
\concept{GroupItem functions}
