% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupItemResource}
\alias{GroupItemResource}
\title{GroupItemResource Object}
\usage{
GroupItemResource(id = NULL)
}
\arguments{
\item{id}{The channel, video, playlist, or asset ID that YouTube uses to uniquely identify the item that is being added to the group}
}
\value{
GroupItemResource object
}
\description{
GroupItemResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GroupItemResource functions}
