% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\name{groups.delete}
\alias{groups.delete}
\title{Deletes a group.}
\usage{
groups.delete(id = NULL, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{id}{The \code{id} parameter specifies the YouTube group ID of the group that is being deleted}

\item{onBehalfOfContentOwner}{This parameter can only be used in a properly authorized request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner, https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/analytics}{Google Documentation}
}
