% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\name{groups.list}
\alias{groups.list}
\title{Returns a collection of groups that match the API request parameters. For example, you can retrieve all groups that the authenticated user owns, or you can retrieve one or more groups by their unique IDs.}
\usage{
groups.list(
  id = NULL,
  mine = NULL,
  pageToken = NULL,
  onBehalfOfContentOwner = NULL
)
}
\arguments{
\item{id}{The \code{id} parameter specifies a comma-separated list of the YouTube group ID(s) for the resource(s) that are being retrieved}

\item{mine}{This parameter can only be used in a properly authorized request}

\item{pageToken}{The \code{pageToken} parameter identifies a specific page in the result set that should be returned}

\item{onBehalfOfContentOwner}{This parameter can only be used in a properly authorized request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner, https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/analytics}{Google Documentation}
}
