% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupSnippet}
\alias{GroupSnippet}
\title{GroupSnippet Object}
\usage{
GroupSnippet(title = NULL, publishedAt = NULL)
}
\arguments{
\item{title}{The group name}

\item{publishedAt}{The date and time that the group was created}
}
\value{
GroupSnippet object
}
\description{
GroupSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group snippet.
}
\concept{GroupSnippet functions}
