% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_functions.R
\name{jobs.reports.list}
\alias{jobs.reports.list}
\title{Lists reports created by a specific job. Returns NOT_FOUND if the job does not exist.}
\usage{
jobs.reports.list(
  jobId,
  createdAfter = NULL,
  startTimeAtOrAfter = NULL,
  pageSize = NULL,
  startTimeBefore = NULL,
  onBehalfOfContentOwner = NULL,
  pageToken = NULL
)
}
\arguments{
\item{jobId}{The ID of the job}

\item{createdAfter}{If set, only reports created after the specified date/time are returned}

\item{startTimeAtOrAfter}{If set, only reports whose start time is greater than or equal the specified date/time are returned}

\item{pageSize}{Requested page size}

\item{startTimeBefore}{If set, only reports whose start time is smaller than the specified date/time are returned}

\item{onBehalfOfContentOwner}{The content owner's external ID on which behalf the user is acting on}

\item{pageToken}{A token identifying a page of results the server should return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reports/}{Google Documentation}
}
