% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{ListGroupItemsResponse}
\alias{ListGroupItemsResponse}
\title{ListGroupItemsResponse Object}
\usage{
ListGroupItemsResponse(items = NULL, errors = NULL, etag = NULL)
}
\arguments{
\item{items}{A list of groups that match the API request parameters}

\item{errors}{Apiary error details}

\item{etag}{The Etag of this resource}
}
\value{
ListGroupItemsResponse object
}
\description{
ListGroupItemsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for GroupsService.ListGroupItems.
}
\concept{ListGroupItemsResponse functions}
