% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{ListGroupsResponse}
\alias{ListGroupsResponse}
\title{ListGroupsResponse Object}
\usage{
ListGroupsResponse(
  nextPageToken = NULL,
  errors = NULL,
  items = NULL,
  etag = NULL
)
}
\arguments{
\item{nextPageToken}{The token that can be used as the value of the \code{pageToken} parameter to retrieve the next page in the result set}

\item{errors}{Apiary error details}

\item{items}{A list of groups that match the API request parameters}

\item{etag}{The Etag of this resource}
}
\value{
ListGroupsResponse object
}
\description{
ListGroupsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for GroupsService.ListGroups.
}
\concept{ListGroupsResponse functions}
