% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{ListReportTypesResponse}
\alias{ListReportTypesResponse}
\title{ListReportTypesResponse Object}
\usage{
ListReportTypesResponse(nextPageToken = NULL, reportTypes = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve next page of results}

\item{reportTypes}{The list of report types}
}
\value{
ListReportTypesResponse object
}
\description{
ListReportTypesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ReportingService.ListReportTypes.
}
\concept{ListReportTypesResponse functions}
