% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_video_metrics.R
\name{playlist_video_metrics}
\alias{playlist_video_metrics}
\title{Get analytics data for all videos in the given playlist.}
\usage{
playlist_video_metrics(
  playlist_id,
  start_date,
  end_date,
  playlist_title = NULL,
  metrics = "views,likes,dislikes",
  dimensions = "day"
)
}
\arguments{
\item{playlist_id}{Character containig the YouTube playlist ID.}

\item{start_date}{Character containing the start date string. "YYYY-MM-DD"}

\item{end_date}{Character containing the end date string. "YYYY-MM-DD"}

\item{playlist_title}{Character playlist title. Optional}

\item{metrics}{Character comma-separated list of metrics to fetch.}

\item{dimensions}{Character comma-separated list of dimensions.}
}
\description{
Get analytics data for all videos in the given playlist.
}
