% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{QueryResponse}
\alias{QueryResponse}
\title{QueryResponse Object}
\usage{
QueryResponse(columnHeaders = NULL, rows = NULL, errors = NULL)
}
\arguments{
\item{columnHeaders}{This value specifies information about the data returned in the \code{rows} fields}

\item{rows}{The list contains all rows of the result table}

\item{errors}{When set, indicates that the operation failed}
}
\value{
QueryResponse object
}
\description{
QueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for TargetedQueriesService.Query.
}
\concept{QueryResponse functions}
