% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\name{reports.query}
\alias{reports.query}
\title{Retrieve your YouTube Analytics reports.}
\usage{
reports.query(
  metrics = NULL,
  filters = NULL,
  includeHistoricalChannelData = NULL,
  currency = NULL,
  ids = NULL,
  endDate = NULL,
  startDate = NULL,
  dimensions = NULL,
  startIndex = NULL,
  sort = NULL,
  maxResults = NULL
)
}
\arguments{
\item{metrics}{A comma-separated list of YouTube Analytics metrics, such as \code{views} or \verb{likes,dislikes}}

\item{filters}{A list of filters that should be applied when retrieving YouTube Analytics data}

\item{includeHistoricalChannelData}{If set to true historical data (i}

\item{currency}{The currency to which financial metrics should be converted}

\item{ids}{Identifies the YouTube channel or content owner for which you are retrieving YouTube Analytics data}

\item{endDate}{The end date for fetching YouTube Analytics data}

\item{startDate}{The start date for fetching YouTube Analytics data}

\item{dimensions}{A comma-separated list of YouTube Analytics dimensions, such as \code{views} or \verb{ageGroup,gender}}

\item{startIndex}{An index of the first entity to retrieve}

\item{sort}{A comma-separated list of dimensions or metrics that determine the sort order for YouTube Analytics data}

\item{maxResults}{The maximum number of rows to include in the response}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner, https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/analytics}{Google Documentation}
}
