% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{ReportType}
\alias{ReportType}
\title{ReportType Object}
\usage{
ReportType(deprecateTime = NULL, name = NULL, systemManaged = NULL, id = NULL)
}
\arguments{
\item{deprecateTime}{The date/time when this report type was/will be deprecated}

\item{name}{The name of the report type (max}

\item{systemManaged}{True if this a system-managed report type; otherwise false}

\item{id}{The ID of the report type (max}
}
\value{
ReportType object
}
\description{
ReportType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A report type.
}
\concept{ReportType functions}
