% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{response_to_data_frame}
\alias{response_to_data_frame}
\title{Internal helpler function to parse the API response to a data.frame
and set the column names appropriately.}
\usage{
response_to_data_frame(r)
}
\arguments{
\item{r}{API response list from gargle::response_process}
}
\value{
data.frame parsed API response.
}
\description{
Internal helpler function to parse the API response to a data.frame
and set the column names appropriately.
}
