% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{ResultTableColumnHeader}
\alias{ResultTableColumnHeader}
\title{YouTube Analytics API Objects
Retrieves your YouTube Analytics data.}
\usage{
ResultTableColumnHeader(dataType = NULL, name = NULL, columnType = NULL)
}
\arguments{
\item{dataType}{The type of the data in the column (\code{STRING}, \code{INTEGER}, \code{FLOAT}, etc}

\item{name}{The name of the dimension or metric}

\item{columnType}{The type of the column (\code{DIMENSION} or \code{METRIC})}
}
\value{
ResultTableColumnHeader object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2022-01-07 12:06:31
filename: /Users/jalvarado/apps/cdc/metrics/r_packages//googleyoutubeAnalyticsv2.auto/R/youtubeAnalytics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ResultTableColumnHeader Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The description of a column of the result table.
}
\concept{ResultTableColumnHeader functions}
