% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{video.demographics}
\alias{video.demographics}
\title{Query the Analytics API and retrieve demographics information}
\usage{
video.demographics(
  video_id,
  start_date,
  end_date,
  filters = NULL,
  sort = "ageGroup,gender",
  ids = "channel==MINE"
)
}
\arguments{
\item{video_id}{YouTube video ID string}

\item{start_date}{date string in 'YYYY-mm-dd' format}

\item{end_date}{date string in 'YYYY-mm-dd' format}

\item{filters}{string of semi-colon separated filters to apply to the query}

\item{sort}{string of comma separated list of dimensions to sort by.  Prepend
the dimension name with '-' to sort descending}

\item{ids}{string of semi-colon separated}
}
\value{
data.frame
}
\description{
Query the Analytics API and retrieve demographics information
}
