% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{video.query}
\alias{video.query}
\title{Query the Analytics API and retrieve metrics for the given video id}
\usage{
video.query(
  video_id,
  metrics = NULL,
  filters = NULL,
  start_date = NULL,
  end_date = NULL,
  dimensions = NULL,
  sort = NULL,
  ids = c("channel==MINE")
)
}
\arguments{
\item{video_id}{YouTube video id string}

\item{metrics}{string of comma separated metric names to query}

\item{filters}{string of semi-colon separated filters to apply to the query}

\item{start_date}{date string in 'YYYY-mm-dd' format}

\item{end_date}{date string in 'YYYY-mm-dd' format}

\item{dimensions}{string of comma separated dimenension names}

\item{sort}{string of comma separated list of dimensions to sort by.  Prepend
the dimension name with '-' to sort descending}

\item{ids}{string of semi-colon separated id terms.  i.e. 'channel==MINE'}
}
\value{
data.frame
}
\description{
Query the Analytics API and retrieve metrics for the given video id
}
