% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{vplaylist.query}
\alias{vplaylist.query}
\title{Take a vector of playlist ids and call playlists.query for each.}
\usage{
vplaylist.query(playlist_ids, ...)
}
\arguments{
\item{playlist_ids}{Vector of playlist id strings}
}
\value{
data.frame containing the results from all queries
}
\description{
Take a vector of playlist ids and call playlists.query for each.
}
