% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{vvideo.demographics}
\alias{vvideo.demographics}
\title{Query the Analytics API and retrieve metrics for the given video ids}
\usage{
vvideo.demographics(video_ids, ...)
}
\arguments{
\item{video_ids}{vector of YouTube video id strings}

\item{filters}{string of semi-colon separated filters to apply to the query}

\item{start_date}{date string in 'YYYY-mm-dd' format}

\item{end_Date}{date string in 'YYYY-mm-dd' format}

\item{sort}{string of comma separated list of dimensions to sort by.  Prepend
the dimension name with '-' to sort descending}

\item{ids}{string of semi-colon separated id terms.  i.e. 'channel==MINE'}
}
\value{
data.frame
}
\description{
Query the Analytics API and retrieve metrics for the given video ids
}
