% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\docType{package}
\name{youtubeAnalytics_googleAuthR}
\alias{youtubeAnalytics_googleAuthR}
\title{YouTube Analytics API
Retrieves your YouTube Analytics data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
at 2022-01-07 12:06:30
filename: /Users/jalvarado/apps/cdc/metrics/r_packages//googleyoutubeAnalyticsv2.auto/R/youtubeAnalytics_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
\item https://www.googleapis.com/auth/youtubepartner
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
\item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.readonly
}
}
