% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_functions.R
\docType{package}
\name{youtubereporting_googleAuthR}
\alias{youtubereporting_googleAuthR}
\title{YouTube Reporting API
Schedules reporting jobs containing your YouTube Analytics data and downloads the resulting bulk data reports in the form of CSV files.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
at 2022-01-14 13:42:58
filename: /Users/jalvarado/apps//googleyoutubereportingv1.auto/R/youtubereporting_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
\item https://www.googleapis.com/auth/yt-analytics.readonly
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
}
}
