% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_analytics_auth}
\alias{yt_analytics_auth}
\title{Authorize YouTube Analytics API}
\usage{
yt_analytics_auth(
  email = gargle::gargle_oauth_email(),
  path = NULL,
  scopes = default_scopes,
  cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default(),
  token = NULL
)
}
\description{
Authorize YouTube Analytics API
}
\concept{auth}
