% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_analytics_auth_configure}
\alias{yt_analytics_auth_configure}
\title{Change the defaults used during the authentication process.}
\usage{
yt_analytics_auth_configure(app, path, api_key)
}
\arguments{
\item{app}{httr::oauth_app}

\item{path}{String path pointing to a JSON file describing an OAuth 2.0 app}

\item{api_key}{String containing an API key that can be used to authenticate
with Google APIs}
}
\description{
Change the defaults used during the authentication process.
}
