% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_analytics_has_token}
\alias{yt_analytics_has_token}
\title{Check if there is currently an OAuth 2.0 token available in the
auth object.}
\usage{
yt_analytics_has_token()
}
\value{
logical
}
\description{
Check if there is currently an OAuth 2.0 token available in the
auth object.
}
