% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_analytics_oauth_app}
\alias{yt_analytics_oauth_app}
\title{Retrieve the current OAuth application from the auth object.}
\usage{
yt_analytics_oauth_app()
}
\value{
httr::oauth_app
}
\description{
Retrieve the current OAuth application from the auth object.
}
