% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_analytics_token}
\alias{yt_analytics_token}
\title{Fetch an authentication token from Google OAuth services if one has not
already been fetched.}
\usage{
yt_analytics_token()
}
\value{
httr::config
}
\description{
Fetch an authentication token from Google OAuth services if one has not
already been fetched.
}
