% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_auth}
\alias{yt_auth}
\title{Authenticate with the YouTube API}
\usage{
yt_auth(
  email = NULL,
  scopes = c("https://www.googleapis.com/auth/youtube",
    "https://www.googleapis.com/auth/youtube.readonly",
    "https://www.googleapis.com/auth/yt-analytics-monetary.readonly",
    "https://www.googleapis.com/auth/yt-analytics.readonly"),
  cache = gargle::gargle_oauth_cache()
)
}
\arguments{
\item{email}{An existing cached email to authenticate with.}

\item{scopes}{Scopes to use in the API requests}

\item{cache}{Where to store the authentication tokens}
}
\value{
an OAuth token object.
}
\description{
Authenticate with the YouTube API
}
