% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{find_playlist}
\alias{find_playlist}
\title{Search YouTube for a playlist}
\usage{
find_playlist(channel_id = NULL, q = NULL, ...)
}
\arguments{
\item{channel_id}{Character containing the YouTube channel ID}

\item{...}{Additional parameters passed to \code{tuber::get_playlists}}
}
\description{
Search YouTube for a playlist
}
\seealso{
tuber::get_playlists
}
