% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_videos.R
\name{get_playlist_videos}
\alias{get_playlist_videos}
\title{Get all videos in the given playlist}
\usage{
get_playlist_videos(playlist_id, ...)
}
\arguments{
\item{playlist_id}{Character containing the YouTube playlist ID.}

\item{...}{Additional arguments passed to \code{tuber::get_playlist_items}}
}
\description{
Get all videos in the given playlist
}
\seealso{
\code{\link[tuber:get_playlist_items]{tuber::get_playlist_items()}} for additional arguments
}
