% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{make_request}
\alias{make_request}
\title{Make an API request to a Google API}
\usage{
make_request(x, ..., user_agent = yt_analytics_ua())
}
\description{
Make an API request to a Google API
}
\concept{requests}
