% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{make_request}
\alias{make_request}
\title{Make an API request to a Google API}
\usage{
make_request(x, ..., user_agent = yt_analytics_ua())
}
\arguments{
\item{x}{List. Holds the components for an HTTP request, presumably created
with \code{\link[gargle:request_develop]{request_develop()}} or \code{\link[gargle:request_build]{request_build()}}. Must contain a \code{method} and
\code{url}. If present, \code{body} and \code{token} are used.}

\item{...}{Optional arguments passed through to the HTTP method. Currently
neither gargle nor httr checks that all are used, so be aware that unused
arguments may be silently ignored.}

\item{user_agent}{A user agent string, prepared by \code{\link[httr:user_agent]{httr::user_agent()}}. When
in doubt, a client package should have an internal function that extends
\code{gargle_user_agent()} by prepending its return value with the client
package's name and version.}
}
\description{
Make an API request to a Google API
}
\concept{requests}
