% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{playlist.demographics}
\alias{playlist.demographics}
\title{Query the Analytics API for playlist demographics metrics}
\usage{
playlist.demographics(playlist_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{playlist_id}{YouTube Playlist ID string}

\item{start_date}{Beginning date of the query date range.}

\item{end_date}{End date of the query date range.}
}
\value{
DataFrame
}
\description{
Query the Analytics API for playlist demographics metrics
}
