% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{playlists.query}
\alias{playlists.query}
\title{Query the analytics API for playlist metrics}
\usage{
playlists.query(
  playlist_id,
  startDate = NULL,
  endDate = NULL,
  metrics = NULL,
  dimensions = NULL,
  sort = NULL,
  filters = NULL,
  ids = NULL
)
}
\description{
Query the analytics API for playlist metrics
}
