% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{video.query}
\alias{video.query}
\title{Query the Analytics API and retrieve metrics for the given video id}
\usage{
video.query(
  video_id,
  metrics = NULL,
  filters = NULL,
  start_date = NULL,
  end_date = NULL,
  dimensions = NULL,
  sort = NULL,
  ids = c("channel==MINE")
)
}
\description{
Query the Analytics API and retrieve metrics for the given video id
}
