% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_auth}
\alias{yt_auth}
\title{Authenticate with the YouTube API}
\usage{
yt_auth()
}
\value{
Invisibly, the token that has been saved to the session
}
\description{
A wrapper for googleAuthR::gar_auth
}
\details{
If you have set the environment variable \code{YT_AUTH_FILE} to a valid file location,
the function will look there for authentication details.
Otherwise it will look in the working directory for hte '.httr-oauth' file, whicih if not present
will trigger an authentication flow via Google login screen in your browser.
}
