% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{yt_auth}
\alias{yt_auth}
\title{Authenticate with the YouTube API}
\usage{
yt_auth(
  email = NULL,
  scopes = c("https://www.googleapis.com/auth/youtube",
    "https://www.googleapis.com/auth/youtube.readonly",
    "https://www.googleapis.com/auth/yt-analytics-monetary.readonly",
    "https://www.googleapis.com/auth/yt-analytics.readonly"),
  cache = gargle::gargle_oauth_cache()
)
}
\arguments{
\item{email}{An existing cached email to authenticate with.}

\item{scopes}{Scopes to use in the API requests}

\item{cache}{Where to store the authentication tokens}
}
\value{
an OAuth token object.
}
\description{
Authenticate with the YouTube API
}
\examples{
\dontrun {
# starts the auth process with default options
yt_auth()

# switch between cached authentication credentials
# by providing an email cache key.
# The first time you use the new scope you will go through the regular
# authentication flow.  Subsequent calls with automatically authenticate.
yt_auth(email = "your@email.com")

# ... query the YouTube Analytics API ...
}

}
